tinyml-modelmaker/pyproject.toml
[build-system]
requires = [ "setuptools>=80",
	"wheel",]
build-backend = "setuptools.build_meta"

[project]
name = "tinyml_modelmaker"
version = "1.2.0"
description = "Tiny ML ModelMaker"
readme = "README.md"
requires-python = "==3.10.*"
classifiers = [ "Development Status :: 4 - Beta",
	"Programming Language :: Python :: 3.10",]
dependencies = [ 
    "defusedxml==0.7.1",
	"numpy==2.2.6",
	"pandas==2.3.0",
	"pathlib2==2.3.7.post1",
	"PyYAML==6.0.2",
	"tqdm==4.67.1",
	"requests==2.32.4",
	"onnx==1.18.0",
	"onnxruntime==1.22.0",
	"protobuf==6.31.1",
	"scipy==1.15.3",
	"tabulate==0.9.0",
	"torcheval==0.0.7",
	"torchinfo==1.8.0",
	"cryptography==45.0.4",
	"matplotlib==3.10.3",
	"scikit-learn==1.7.0",
	"torch==2.7.1",
    "tinyml_tinyverse @ git+https://github.com/TexasInstruments/tinyml-tensorlab.git@r1.2#subdirectory=tinyml-tinyverse",
	"tinyml_torchmodelopt @ git+https://github.com/TexasInstruments/tinyml-tensorlab.git@r1.2#subdirectory=tinyml-modeloptimization/torchmodelopt",
	"ti_mcu_nnc @ https://software-dl.ti.com/mctools/esd/tvm/mcu/ti_mcu_nnc-2.1.0-cp310-cp310-linux_x86_64.whl ; platform_system == 'Linux'",
	"ti_mcu_nnc @ https://software-dl.ti.com/mctools/esd/tvm/mcu/ti_mcu_nnc-2.1.0-cp310-cp310-win_amd64.whl ; platform_system == 'Windows'",
	"Cython==3.1.2",
	"protobuf==6.31.1",
	"onnxsim==0.4.36",
	"onnxoptimizer==0.3.13",
	"progressbar==2.5",
	"rich==14.0.0",]
authors = [{name = "Adithya Thonse", email = "thonse@ti.com"}]

[project.license]
text = "BSD 3-Clause"

[project.urls]
Homepage = "https://github.com/TexasInstruments/tinyml-tensorlab/tree/main/tinyml-modelmaker"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = [ "tinyml_modelmaker*",]
exclude = [ "tests*",]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-cov>=5.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.dynamic.version]
attr = "tinyml_modelmaker.__version__"
